refactored code and trying to fricking fix the ores

+++ src/main/java/io/github/vampirestudios/raa/generation/chunkgenerator/FlatCavesChunkGenerator.java
@@ -1,16 +1,9 @@
 package io.github.vampirestudios.raa.generation.chunkgenerator;
 
-import net.minecraft.block.Blocks;
-import net.minecraft.entity.EntityCategory;
-import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
-import net.minecraft.world.biome.Biome;
 import net.minecraft.world.biome.source.BiomeSource;
 import net.minecraft.world.gen.chunk.CavesChunkGeneratorConfig;
 import net.minecraft.world.gen.chunk.SurfaceChunkGenerator;
-import net.minecraft.world.gen.feature.Feature;
-
-import java.util.List;
 
 public class FlatCavesChunkGenerator extends SurfaceChunkGenerator<CavesChunkGeneratorConfig> {
     private final double[] noiseFalloff = this.buildNoiseFalloff();
@@ -34,11 +27,11 @@ public class FlatCavesChunkGenerator extends SurfaceChunkGenerator<CavesChunkGen
     private double[] buildNoiseFalloff() {
         double[] doubles_1 = new double[this.getNoiseSizeY()];
 
-        for(int int_1 = 0; int_1 < this.getNoiseSizeY(); ++int_1) {
-            doubles_1[int_1] = Math.cos((double)int_1 * 3.141592653589793D * 6.0D / (double)this.getNoiseSizeY())* 1.1;
-            double double_1 = (double)int_1;
+        for (int int_1 = 0; int_1 < this.getNoiseSizeY(); ++int_1) {
+            doubles_1[int_1] = Math.cos((double) int_1 * 3.141592653589793D * 6.0D / (double) this.getNoiseSizeY()) * 1.1;
+            double double_1 = int_1;
             if (int_1 > this.getNoiseSizeY() / 2) {
-                double_1 = (double)(this.getNoiseSizeY() - int_1);
+                double_1 = this.getNoiseSizeY() - int_1;
             }
 
             if (double_1 < 4.0D) {
